requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
python-jose[cryptography]
passlib[bcrypt]
python-multipart
email-validator
psycopg2-binary
python-dotenv
